vold: set dirty ratio to 20 when UMS is active to improve SD io performance
The default value of dirty ratio is 20.In order to improves UI responsiveness
when copying large amount of data to the device, google android developer
set dirty ratio to 0 when UMS is active. But it will lead to very slowly
ums write speed on our hardware platform. So we set dirty ratio to 20 when
UMS is active to improve SD io performance.

Signed-off-by: Royce Jiang <jiangxd@embedinfo.com>